Decouple validator scope registry from naming host
